+++ .eslintignore
@@ -1,2 +1,3 @@
 dist/
 jest.config.ts
+interfaces/
