+++ .eslintrc.json
@@ -29,9 +29,9 @@
     "import/resolver": {
       "typescript": {},
       "alias": [
-        ["@controllers", "./src/controllers"],
+        ["@controllers", "./src/infra/http/controllers"],
         ["@DTOs", "./src/DTOs"],
-        ["@data-access", "./src/data-access"],
+        ["@prisma-repositories", "./src/infra/data-access/prisma/repositories"],
         ["@database", "./src/lib/prisma/db.ts"],
         ["@use-cases", "./src/use-cases"]
       ]
