+++ tsconfig.json
@@ -18,9 +18,9 @@
     "baseUrl": "./",
     "strict": true,
     "paths": {
-      "@controllers": ["src/http/controllers"],
+      "@controllers": ["src/infra/http/controllers"],
       "@DTOs": ["./src/DTOs"],
-      "@data-access": ["src/data-access"],
+      "@prisma-repositories": ["src/infra/data-access/prisma/repositories"],
       "@database": ["src/lib/prisma/db.ts"],
       "@use-cases": ["src/use-cases"]
     }
